==> Notebooks/04_nlp_pipeline.ipynb <==
import pandas as pd
import os
import re
from datetime import datetime

fn_list = os.listdir("data/after-nlp-pipeline")

#after-scraping-08-29-2019-01-15-09.csv
dates = list(map(lambda x: {'fileName': x, 'date': re.search("([0-9]{2}\-[0-9]{2}\-[0-9]{4}-[0-9]{2}-[0-9]{2}-[0-9]{2})", x).group()}, fn_list)) 
dates = list(map(lambda x: {'fileName': x['fileName'], 'date': datetime.strptime(x['date'], "%m-%d-%Y-%H-%M-%S")}, dates))
dates.sort(key=lambda x: x['date'], reverse=True)

fn_newest = dates[0]['fileName']

f = open("data/after-nlp-pipeline/{}".format(fn_newest), 'r', encoding="utf-8")
df = pd.read_csv(f, index_col=0, converters={'text_scraped_words': lambda x: x[1:-1].replace("'", "").split(', ')})
f.close()

print("Successfully imported file: {}".format(fn_newest)) 

df
